Support minimize function of multi-window
When minimize icon is clicked, we will use Activity.moveTaskToBack
method to move Activity's task to back based on z-order.

After minimzed, the launcher will launch the origin task of app, and
move it to front for many occasion.

TODO: Fix minimized function failed when Activity is in maximized mode
in some occasion.

Change-Id: I04169697b5f40bebd5814e7def756b82a974d160
Signed-off-by: utzcoz <utzcoz@outlook.com>